+++ aiocamedomotic/utils.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Constants for the CAME Domotic API."""
+"""Utilities for the CAME Domotic API."""
 
 import logging
